Update: init.sh

+++ init.sh
@@ -3,7 +3,7 @@
 if [ "$(uname)" == "Darwin" ] ; then
 	.bin/macos.sh
 	# Amazon Q setting file
-	ln -snfv ${PWD}/amazon-q_settings.json "${HOME}/Library/Application Support/amazon-q/settings.json"
+	[[ -f "/Applications/Amazon Q.app" ]] && ln -snfv ${PWD}/amazon-q_settings.json "${HOME}/Library/Application Support/amazon-q/settings.json"
 fi
 
 # symlink
